+++ ddtrace/contrib/graphql/__init__.py
@@ -51,8 +51,12 @@ required_modules = ["graphql"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.graphql.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.graphql.patch import get_version
+        from ..internal.graphql.patch import patch
+        from ..internal.graphql.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
