+++ ddtrace/contrib/grpc/__init__.py
@@ -82,8 +82,12 @@ required_modules = ["grpc"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.grpc.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.grpc.patch import get_version
+        from ..internal.grpc.patch import patch
+        from ..internal.grpc.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
